give ColorGradient it's own module
